+++ CHANGELOG.md
@@ -1,17 +1,28 @@
 # CHANGELOG
 
-## V2.0.2
+## 3.0.0
+
+- Add support for GitHub Apps
+- Drop support for PHP 8.0 and 8.1
+- Drop support for Laravel 8 and 9
+- Drop support for GitHub Personal Access Tokens
+
+Run `php artisan vendor:publish --tag=publish-config` to publish the new configuration file.
+
+See "GitHub credentials" in README.md for information about creating a GitHub App.
+
+## 2.0.2
 
 - Fix cms navigation bug
 
-## V2.0.1
+## 2.0.1
 
 - Add event for when publication was started
 
-## V2.0.0
+## 2.0.0
 
 - Nova 4 added as dependency. The tool is now compatible with Nova 4.
 
-## V1.0.0
+## 1.0.0
 
 - Initial version for Nova 3.